project: move SDL_rtf.h into SDL3_rtf subfolder
